Rename MCP to plan-nyc-mich and curate tools to NYC↔Michigan routes
Moves mcp/ → mcp/plan-nyc-mich/ and replaces the generic flight_price /
price_matrix tools with two route-curated tools:

  - weekend_price_matrix: cheapest one-way price per weekend day across
    JFK/LGA/NYC × GRR/DTW for the next N weekends.
  - cheapest_weekend_trip: single best (route, date) pair from that set.

Going forward we add a new MCP per route family rather than parameterizing
this one.
